requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
frozenlist==1.2.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
multidict==5.2.0
oauth2client==4.1.3
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
